refactor: move system table code to separate module
