+++ influxdb3_server/src/query_executor.rs
@@ -1,13 +1,10 @@
 //! module for query executor
+use crate::system_tables::{SystemSchemaProvider, SYSTEM_SCHEMA};
 use crate::{QueryExecutor, QueryKind};
-use arrow::array::{
-    ArrayRef, BooleanArray, DurationNanosecondArray, GenericListBuilder, Int64Array, Int64Builder,
-    StringBuilder, StructArray, TimestampNanosecondArray,
-};
+use arrow::array::{ArrayRef, Int64Builder, StringBuilder, StructArray};
 use arrow::datatypes::SchemaRef;
 use arrow::record_batch::RecordBatch;
-use arrow_array::UInt64Array;
-use arrow_schema::{ArrowError, TimeUnit};
+use arrow_schema::ArrowError;
 use async_trait::async_trait;
 use data_types::NamespaceId;
 use datafusion::catalog::schema::SchemaProvider;
@@ -23,7 +20,7 @@ use datafusion::physical_plan::ExecutionPlan;
 use datafusion::prelude::Expr;
 use datafusion_util::config::DEFAULT_SCHEMA;
 use datafusion_util::MemoryStream;
-use influxdb3_write::last_cache::{LastCacheFunction, LastCacheInfo, LastCacheProvider};
+use influxdb3_write::last_cache::LastCacheFunction;
 use influxdb3_write::{
     catalog::{Catalog, DatabaseSchema},
     WriteBuffer,
@@ -31,15 +28,14 @@ use influxdb3_write::{
 use iox_query::exec::{Executor, IOxSessionContext, QueryConfig};
 use iox_query::frontend::sql::SqlQueryPlanner;
 use iox_query::provider::ProviderBuilder;
+use iox_query::query_log::QueryLog;
+use iox_query::query_log::QueryText;
 use iox_query::query_log::StateReceived;
 use iox_query::query_log::{QueryCompletedToken, QueryLogEntries};
-use iox_query::query_log::{QueryLog, QueryLogEntryState};
-use iox_query::query_log::{QueryPhase, QueryText};
 use iox_query::QueryDatabase;
 use iox_query::{QueryChunk, QueryNamespace};
 use iox_query_influxql::frontend::planner::InfluxQLQueryPlanner;
 use iox_query_params::StatementParams;
-use iox_system_tables::{IoxSystemTable, SystemTableProvider};
 use metric::Registry;
 use observability_deps::tracing::{debug, info};
 use schema::Schema;
@@ -575,396 +571,4 @@ impl<B: WriteBuffer> TableProvider for QueryTable<B> {
 
         provider.scan(ctx, projection, &filters, limit).await
     }
-}
-
-pub const SYSTEM_SCHEMA: &str = "system";
-
-const QUERIES_TABLE: &str = "queries";
-const LAST_CACHES_TABLE: &str = "last_caches";
-const _PARQUET_FILES_TABLE: &str = "parquet_files";
-
-struct SystemSchemaProvider {
-    tables: HashMap<&'static str, Arc<dyn TableProvider>>,
-}
-
-impl std::fmt::Debug for SystemSchemaProvider {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        let mut keys = self.tables.keys().copied().collect::<Vec<_>>();
-        keys.sort_unstable();
-
-        f.debug_struct("SystemSchemaProvider")
-            .field("tables", &keys.join(", "))
-            .finish()
-    }
-}
-
-impl SystemSchemaProvider {
-    fn new(
-        db_name: impl Into<String>,
-        _catalog: Arc<Catalog>,
-        query_log: Arc<QueryLog>,
-        last_cache_provider: Arc<LastCacheProvider>,
-    ) -> Self {
-        let mut tables = HashMap::<&'static str, Arc<dyn TableProvider>>::new();
-        let queries = Arc::new(SystemTableProvider::new(Arc::new(QueriesTable::new(
-            query_log,
-        ))));
-        tables.insert(QUERIES_TABLE, queries);
-        let last_caches = Arc::new(SystemTableProvider::new(Arc::new(LastCachesTable::new(
-            db_name,
-            last_cache_provider,
-        ))));
-        tables.insert(LAST_CACHES_TABLE, last_caches);
-        Self { tables }
-    }
-}
-
-#[async_trait]
-impl SchemaProvider for SystemSchemaProvider {
-    fn as_any(&self) -> &dyn Any {
-        self as &dyn Any
-    }
-
-    fn table_names(&self) -> Vec<String> {
-        let mut names = self
-            .tables
-            .keys()
-            .map(|s| (*s).to_owned())
-            .collect::<Vec<_>>();
-        names.sort();
-        names
-    }
-
-    async fn table(&self, name: &str) -> Result<Option<Arc<dyn TableProvider>>, DataFusionError> {
-        Ok(self.tables.get(name).cloned())
-    }
-
-    fn table_exist(&self, name: &str) -> bool {
-        self.tables.contains_key(name)
-    }
-}
-
-struct QueriesTable {
-    schema: SchemaRef,
-    query_log: Arc<QueryLog>,
-}
-
-impl QueriesTable {
-    fn new(query_log: Arc<QueryLog>) -> Self {
-        Self {
-            schema: queries_schema(),
-            query_log,
-        }
-    }
-}
-
-#[async_trait::async_trait]
-impl IoxSystemTable for QueriesTable {
-    fn schema(&self) -> SchemaRef {
-        Arc::clone(&self.schema)
-    }
-
-    async fn scan(
-        &self,
-        _filters: Option<Vec<Expr>>,
-        _limit: Option<usize>,
-    ) -> Result<RecordBatch, DataFusionError> {
-        let schema = self.schema();
-
-        let entries = self
-            .query_log
-            .entries()
-            .entries
-            .into_iter()
-            .map(|e| e.state())
-            .collect::<Vec<_>>();
-
-        from_query_log_entries(Arc::clone(&schema), &entries)
-    }
-}
-
-fn queries_schema() -> SchemaRef {
-    let columns = vec![
-        Field::new("id", DataType::Utf8, false),
-        Field::new("phase", DataType::Utf8, false),
-        Field::new(
-            "issue_time",
-            DataType::Timestamp(TimeUnit::Nanosecond, None),
-            false,
-        ),
-        Field::new("query_type", DataType::Utf8, false),
-        Field::new("query_text", DataType::Utf8, false),
-        Field::new("partitions", DataType::Int64, true),
-        Field::new("parquet_files", DataType::Int64, true),
-        Field::new(
-            "plan_duration",
-            DataType::Duration(TimeUnit::Nanosecond),
-            true,
-        ),
-        Field::new(
-            "permit_duration",
-            DataType::Duration(TimeUnit::Nanosecond),
-            true,
-        ),
-        Field::new(
-            "execute_duration",
-            DataType::Duration(TimeUnit::Nanosecond),
-            true,
-        ),
-        Field::new(
-            "end2end_duration",
-            DataType::Duration(TimeUnit::Nanosecond),
-            true,
-        ),
-        Field::new(
-            "compute_duration",
-            DataType::Duration(TimeUnit::Nanosecond),
-            true,
-        ),
-        Field::new("max_memory", DataType::Int64, true),
-        Field::new("success", DataType::Boolean, false),
-        Field::new("running", DataType::Boolean, false),
-        Field::new("cancelled", DataType::Boolean, false),
-        Field::new("trace_id", DataType::Utf8, true),
-    ];
-
-    Arc::new(DatafusionSchema::new(columns))
-}
-
-fn from_query_log_entries(
-    schema: SchemaRef,
-    entries: &[Arc<QueryLogEntryState>],
-) -> Result<RecordBatch, DataFusionError> {
-    let mut columns: Vec<ArrayRef> = vec![];
-
-    columns.push(Arc::new(
-        entries
-            .iter()
-            .map(|e| Some(e.id.to_string()))
-            .collect::<StringArray>(),
-    ));
-
-    columns.push(Arc::new(
-        entries
-            .iter()
-            .map(|e| Some(e.phase.name()))
-            .collect::<StringArray>(),
-    ));
-
-    columns.push(Arc::new(
-        entries
-            .iter()
-            .map(|e| e.issue_time)
-            .map(|ts| Some(ts.timestamp_nanos()))
-            .collect::<TimestampNanosecondArray>(),
-    ));
-
-    columns.push(Arc::new(
-        entries
-            .iter()
-            .map(|e| Some(&e.query_type))
-            .collect::<StringArray>(),
-    ));
-
-    columns.push(Arc::new(
-        entries
-            .iter()
-            .map(|e| Some(e.query_text.to_string()))
-            .collect::<StringArray>(),
-    ));
-
-    columns.push(Arc::new(
-        entries.iter().map(|e| e.partitions).collect::<Int64Array>(),
-    ));
-
-    columns.push(Arc::new(
-        entries
-            .iter()
-            .map(|e| e.parquet_files)
-            .collect::<Int64Array>(),
-    ));
-
-    columns.push(Arc::new(
-        entries
-            .iter()
-            .map(|e| e.plan_duration.map(|d| d.as_nanos() as i64))
-            .collect::<DurationNanosecondArray>(),
-    ));
-
-    columns.push(Arc::new(
-        entries
-            .iter()
-            .map(|e| e.permit_duration.map(|d| d.as_nanos() as i64))
-            .collect::<DurationNanosecondArray>(),
-    ));
-
-    columns.push(Arc::new(
-        entries
-            .iter()
-            .map(|e| e.execute_duration.map(|d| d.as_nanos() as i64))
-            .collect::<DurationNanosecondArray>(),
-    ));
-
-    columns.push(Arc::new(
-        entries
-            .iter()
-            .map(|e| e.end2end_duration.map(|d| d.as_nanos() as i64))
-            .collect::<DurationNanosecondArray>(),
-    ));
-
-    columns.push(Arc::new(
-        entries
-            .iter()
-            .map(|e| e.compute_duration.map(|d| d.as_nanos() as i64))
-            .collect::<DurationNanosecondArray>(),
-    ));
-
-    columns.push(Arc::new(
-        entries.iter().map(|e| e.max_memory).collect::<Int64Array>(),
-    ));
-
-    columns.push(Arc::new(
-        entries
-            .iter()
-            .map(|e| Some(e.success))
-            .collect::<BooleanArray>(),
-    ));
-
-    columns.push(Arc::new(
-        entries
-            .iter()
-            .map(|e| Some(e.running))
-            .collect::<BooleanArray>(),
-    ));
-
-    columns.push(Arc::new(
-        entries
-            .iter()
-            .map(|e| Some(e.phase == QueryPhase::Cancel))
-            .collect::<BooleanArray>(),
-    ));
-
-    columns.push(Arc::new(
-        entries
-            .iter()
-            .map(|e| e.trace_id.map(|x| format!("{:x}", x.0)))
-            .collect::<StringArray>(),
-    ));
-
-    let batch = RecordBatch::try_new(schema, columns)?;
-    Ok(batch)
-}
-
-struct LastCachesTable {
-    db_name: String,
-    schema: SchemaRef,
-    provider: Arc<LastCacheProvider>,
-}
-
-impl LastCachesTable {
-    fn new(db: impl Into<String>, provider: Arc<LastCacheProvider>) -> Self {
-        Self {
-            db_name: db.into(),
-            schema: last_caches_schema(),
-            provider,
-        }
-    }
-}
-
-fn last_caches_schema() -> SchemaRef {
-    let columns = vec![
-        Field::new("table", DataType::Utf8, false),
-        Field::new("name", DataType::Utf8, false),
-        Field::new(
-            "key_columns",
-            DataType::List(Arc::new(Field::new("item", DataType::Utf8, true))),
-            false,
-        ),
-        Field::new(
-            "value_columns",
-            DataType::List(Arc::new(Field::new("item", DataType::Utf8, true))),
-            false,
-        ),
-        Field::new("count", DataType::UInt64, false),
-        Field::new("ttl", DataType::UInt64, false),
-    ];
-    Arc::new(DatafusionSchema::new(columns))
-}
-
-#[async_trait::async_trait]
-impl IoxSystemTable for LastCachesTable {
-    fn schema(&self) -> SchemaRef {
-        Arc::clone(&self.schema)
-    }
-
-    async fn scan(
-        &self,
-        _filters: Option<Vec<Expr>>,
-        _limit: Option<usize>,
-    ) -> Result<RecordBatch, DataFusionError> {
-        let caches = self.provider.get_last_caches_for_db(&self.db_name);
-        from_last_cache_infos(self.schema(), &caches)
-    }
-}
-
-fn from_last_cache_infos(
-    schema: SchemaRef,
-    caches: &[LastCacheInfo],
-) -> Result<RecordBatch, DataFusionError> {
-    let mut columns: Vec<ArrayRef> = vec![];
-
-    // Table Name
-    columns.push(Arc::new(
-        caches
-            .iter()
-            .map(|c| Some(c.table.to_string()))
-            .collect::<StringArray>(),
-    ));
-    // Cache Name
-    columns.push(Arc::new(
-        caches
-            .iter()
-            .map(|c| Some(c.name.to_string()))
-            .collect::<StringArray>(),
-    ));
-    // Key Columns
-    columns.push({
-        let values_builder = StringBuilder::new();
-        let mut builder = GenericListBuilder::<i32, _>::new(values_builder);
-
-        for c in caches {
-            c.key_columns
-                .iter()
-                .for_each(|k| builder.values().append_value(k));
-            builder.append(true);
-        }
-        Arc::new(builder.finish())
-    });
-    // Value Columns
-    columns.push({
-        let values_builder = StringBuilder::new();
-        let mut builder = GenericListBuilder::<i32, _>::new(values_builder);
-
-        for c in caches {
-            c.value_columns
-                .iter()
-                .for_each(|v| builder.values().append_value(v));
-            builder.append(true);
-        }
-        Arc::new(builder.finish())
-    });
-    columns.push(Arc::new(
-        caches
-            .iter()
-            .map(|e| Some(e.count as u64))
-            .collect::<UInt64Array>(),
-    ));
-    columns.push(Arc::new(
-        caches
-            .iter()
-            .map(|e| Some(e.ttl.as_secs()))
-            .collect::<UInt64Array>(),
-    ));
-
-    Ok(RecordBatch::try_new(schema, columns)?)
 }
